requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
torch==2.0.1
matplotlib==3.7.2
seaborn==0.12.2
jupyter==1.0.0
pyyaml==6.0.1
tqdm==4.65.0
pytest==7.4.0
scapy==2.5.0